sentwise model run

+++ data/utils.py
@@ -195,22 +195,22 @@ class InputSentenceCATSDatasetBuilder:
                 p1mat = self.paravecs_npy[p1index_dat[0]:p1index_dat[0] + p1index_dat[1]]
                 p1vec_len = p1mat.shape[0]
                 if p1vec_len < self.max_seq_len:
-                    valid_bits = np.array([1.0] * p1vec_len + [0.0] * (self.max_seq_len - p1vec_len)).reshape((-1,1)).float()
+                    valid_bits = np.array([1.0] * p1vec_len + [0.0] * (self.max_seq_len - p1vec_len)).reshape((-1,1))
                     z = np.zeros((self.max_seq_len - p1vec_len, self.emb_len))
                     p1mat = np.hstack((np.vstack((p1mat, z)), valid_bits))
                 else:
-                    valid_bits = np.array([1.0] * self.max_seq_len).reshape((-1,1)).float()
+                    valid_bits = np.array([1.0] * self.max_seq_len).reshape((-1,1))
                     p1mat = np.hstack((p1mat[:self.max_seq_len], valid_bits))
 
                 p2index_dat = self.paraids_dict[pid2]
                 p2mat = self.paravecs_npy[p2index_dat[0]:p2index_dat[0] + p2index_dat[1]]
                 p2vec_len = p2mat.shape[0]
                 if p2vec_len < self.max_seq_len:
-                    valid_bits = np.array([1.0] * p2vec_len + [0.0] * (self.max_seq_len - p2vec_len)).reshape((-1,1)).float()
+                    valid_bits = np.array([1.0] * p2vec_len + [0.0] * (self.max_seq_len - p2vec_len)).reshape((-1,1))
                     p2mat = np.hstack(
                         (np.vstack((p2mat, np.zeros((self.max_seq_len - p2vec_len, self.emb_len)))), valid_bits))
                 else:
-                    valid_bits = np.array([1.0] * self.max_seq_len).reshape((-1,1)).float()
+                    valid_bits = np.array([1.0] * self.max_seq_len).reshape((-1,1))
                     p2mat = np.hstack((p2mat[:self.max_seq_len], valid_bits))
                 dat_mat = np.hstack((qmat, p1mat, p2mat))
